\name{BM_bernoulli_covariates_fast}
\alias{BM_bernoulli_covariates_fast}
\alias{\S4method{new}{BM_bernoulli_covariates_fast}}
\title{Perform estimation on blockmodels for bernoulli probability distribution
aith covariates}
\description{
    With the provided network and blockmodel type, estimate number of groups,
parameters and node membership, and impact vector of covariates.}
\usage{\S4method{new}{BM_bernoulli_covariates_fast}(MEMBERSHIP_ARG, SCALAR_MODEL_ARG, COVARIATES_ARG, MODEL_ARG)}
\arguments{
    MEMBERSHIP_ARG_ITEM
    SCALAR_MODEL_ARG_ITEM
    COVARIATES_ARG_ITEM
    MODEL_ARG_ITEM
}
