\name{BM_bernoulli_multiplex}
\alias{BM_bernoulli_multiplex}
\alias{\S4method{BM_bernoulli_multiplex}{new}}
\title{Perform estimation on blockmodels for multiplex binary networks}
\description{
    With the provided network and blockmodel type, estimate number of groups,
parameters and node membership}
\usage{\S4method{BM_bernoulli_multiplex}{new}(
    MEMBERSHIP_ARG, 
    MULTIVARIATE_MODEL_ARG, 
    MODEL_ARG)}
\arguments{
    MEMBERSHIP_ARG_ITEM
    MULTIVARIATE_MODEL_ARG_ITEM
    MODEL_ARG_ITEM
}
