\name{BM_gaussian}
\alias{BM_gaussian}
\alias{\S4method{new}{BM_gaussian}}
\title{Perform estimation on blockmodels for gaussian probability distribution}
\description{
    With the provided network and blockmodel type, estimate number of groups,
parameters and node membership}
\usage{\S4method{new}{BM_gaussian}(MEMBERSHIP_ARG, SCALAR_MODEL_ARG, MODEL_ARG)}
\arguments{
    MEMBERSHIP_ARG_ITEM
    SCALAR_MODEL_ARG_ITEM
    MODEL_ARG_ITEM
}
