\name{BM_gaussian_covariates}
\alias{BM_gaussian_covariates}
\alias{\S4method{new}{BM_gaussian_covariates}}
\title{Perform estimation on blockmodels for gaussian probability distribution
with covariates}
\description{
    With the provided network and blockmodel type, estimate number of groups,
parameters and node membership and impact vector of covariates}
\usage{\S4method{new}{BM_gaussian_covariates}(MEMBERSHIP_ARG, SCALAR_MODEL_ARG, COVARIATES_ARG, MODEL_ARG)}
\arguments{
    MEMBERSHIP_ARG_ITEM
    SCALAR_MODEL_ARG_ITEM
    COVARIATES_ARG_ITEM
    MODEL_ARG_ITEM
}
