\name{BM_gaussian_multivariate}
\alias{BM_gaussian_multivariate}
\alias{\S4method{new}{BM_gaussian_multivariate}}
\title{Perform estimation on blockmodels for multivariate gaussian probability distribution}
\description{
    With the provided network and blockmodel type, estimate number of groups,
parameters and node membership}
\usage{\S4method{new}{BM_gaussian_multivariate}(MEMBERSHIP_ARG, MULTIVARIATE_MODEL_ARG, MODEL_ARG)}
\arguments{
    MEMBERSHIP_ARG_ITEM
    MULTIVARIATE_MODEL_ARG_ITEM
    MODEL_ARG_ITEM
}
