\name{BM_poisson_covariates}
\alias{BM_poisson_covariates}
\alias{\S4method{BM_poisson_covariates}{new}}
\title{Perform estimation on blockmodels for poisson probability distribution
aith covariates}
\description{
    With the provided network and blockmodel type, estimate number of groups,
parameters and node membership, and impact vector of covariates}
\usage{\S4method{BM_poisson_covariates}{new}(
    MEMBERSHIP_ARG, 
    SCALAR_MODEL_ARG, 
    COVARIATES_ARG, 
    MODEL_ARG)}
\arguments{
    MEMBERSHIP_ARG_ITEM
    SCALAR_MODEL_ARG_ITEM
    COVARIATES_ARG_ITEM
    MODEL_ARG_ITEM
}
