\name{BM_gaussian}
\alias{BM_gaussian}
\alias{\S4method{BM_gaussian}{new}}
\title{Perform estimation on blockmodels for gaussian probability distribution}
\description{
    With the provided network and blockmodel type, estimate number of groups,
parameters and node membership}
\usage{\S4method{BM_gaussian}{new}(
    MEMBERSHIP_ARG, 
    SCALAR_MODEL_ARG, 
    MODEL_ARG)}
\arguments{
    MEMBERSHIP_ARG_ITEM
    SCALAR_MODEL_ARG_ITEM
    MODEL_ARG_ITEM
}
