\name{BM_gaussian_multivariate}
\alias{BM_gaussian_multivariate}
\alias{\S4method{BM_gaussian_multivariate}{new}}
\title{Perform estimation on blockmodels for multivariate gaussian probability distribution}
\description{
    With the provided network and blockmodel type, estimate number of groups,
parameters and node membership}
\usage{\S4method{BM_gaussian_multivariate}{new}(
    MEMBERSHIP_ARG, 
    MULTIVARIATE_MODEL_ARG, 
    MODEL_ARG)}
\arguments{
    MEMBERSHIP_ARG_ITEM
    MULTIVARIATE_MODEL_ARG_ITEM
    MODEL_ARG_ITEM
}
