\name{BM_poisson}
\alias{BM_poisson}
\alias{\S4method{BM_poisson}{new}}
\title{Perform estimation on blockmodels for poisson probability distribution}
\description{
    With the provided network and blockmodel type, estimate number of groups,
parameters and node membership}
\usage{\S4method{BM_poisson}{new}(
    membership_type, 
    SCALAR_MODEL_ARG, 
    MODEL_ARG)}
\arguments{
    MEMBERSHIP_ARG_ITEM
    SCALAR_MODEL_ARG_ITEM
    MODEL_ARG_ITEM
}
